requirements.txt
fastapi
uvicorn[standard]
pymodbus
pyyaml
